requirements.txt
loguru==0.7.2
matplotlib==3.9.2
numpy==2.1.3
pandas==2.2.2
python-dotenv==1.0.1